# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/me":{"get":{"tags":["users"],"summary":"Get Current User","operationId":"get_current_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users":{"post":{"tags":["users"],"summary":"Add User","operationId":"add_user_users_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationDTO"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["users"],"summary":"Delete User","operationId":"delete_user_users_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/UserDTO"},{"type":"null"}],"title":"Response Delete User Users Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chats":{"get":{"tags":["chats"],"summary":"Get Chats","operationId":"get_chats_chats_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"type":"array","items":{"$ref":"#/components/schemas/Chat"}},{"type":"null"}],"title":"Response Get Chats Chats Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["chats"],"summary":"Add Chat","operationId":"add_chat_chats_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"title","in":"query","required":true,"schema":{"type":"string","title":"Title"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"},"title":"Users"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["chats"],"summary":"Delete Chat","operationId":"delete_chat_chats_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chat_id","in":"query","required":true,"schema":{"type":"integer","title":"Chat Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/Chat"},{"type":"null"}],"title":"Response Delete Chat Chats Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chats/{chatId}":{"get":{"tags":["chats"],"summary":"Get Chat","operationId":"get_chat_chats__chatId__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"integer","title":"Chatid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/ChatView"},{"type":"null"}],"title":"Response Get Chat Chats  Chatid  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chats/{chatId}/documents":{"post":{"tags":["chats"],"summary":"Add Document To Chat","operationId":"add_document_to_chat_chats__chatId__documents_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"integer","title":"Chatid"}},{"name":"documentId","in":"query","required":true,"schema":{"type":"integer","title":"Documentid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["chats"],"summary":"Get Documents In Chat","operationId":"get_documents_in_chat_chats__chatId__documents_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"integer","title":"Chatid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Document"},"title":"Response Get Documents In Chat Chats  Chatid  Documents Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["chats"],"summary":"Delete Documents In Chat","operationId":"delete_documents_in_chat_chats__chatId__documents_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"integer","title":"Chatid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages":{"get":{"tags":["messages"],"summary":"Get Messages","operationId":"get_messages_messages_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chat_id","in":"query","required":true,"schema":{"type":"integer","title":"Chat Id"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"type":"array","items":{"$ref":"#/components/schemas/Message"}},{"type":"null"}],"title":"Response Get Messages Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["messages"],"summary":"Add Message","operationId":"add_message_messages_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"chat_id","in":"query","required":true,"schema":{"type":"integer","title":"Chat Id"}},{"name":"content","in":"query","required":true,"schema":{"type":"string","title":"Content"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["messages"],"summary":"Delete Message","operationId":"delete_message_messages_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"message_id","in":"query","required":true,"schema":{"type":"integer","title":"Message Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/Message"},{"type":"null"}],"title":"Response Delete Message Messages Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate":{"post":{"tags":["tasks"],"summary":"Add Task","operationId":"add_task_generate_post","parameters":[{"name":"prompt","in":"query","required":true,"schema":{"type":"string","title":"Prompt"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["tasks"],"summary":"Get Task","operationId":"get_task_generate_get","parameters":[{"name":"task_id","in":"query","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/add_document_to_index":{"post":{"tags":["tasks"],"summary":"Add Task Add Document To Index","operationId":"add_task_add_document_to_index_add_document_to_index_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"document_id","in":"query","required":true,"schema":{"type":"integer","title":"Document Id"}},{"name":"chat_id","in":"query","required":true,"schema":{"type":"integer","title":"Chat Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["tasks"],"summary":"Get Task Add Document To Index","operationId":"get_task_add_document_to_index_add_document_to_index_get","parameters":[{"name":"task_id","in":"query","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query":{"post":{"tags":["tasks"],"summary":"Add Task Query","operationId":"add_task_query_query_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["tasks"],"summary":"Get Task Query","operationId":"get_task_query_query_get","parameters":[{"name":"task_id","in":"query","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/upload":{"post":{"tags":["documents"],"summary":"Upload","operationId":"upload_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_documents_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/documents":{"get":{"tags":["documents"],"summary":"Get Documents","operationId":"get_documents_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_for_access_token_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_token_post"},"Body_upload_documents_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_documents_upload_post"},"Chat":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"title":{"type":"string","maxLength":256,"title":"Title"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["title"],"title":"Chat"},"ChatView":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"user_ids":{"items":{"type":"integer"},"type":"array","title":"User Ids","default":[]},"messages_ids":{"items":{"type":"integer"},"type":"array","title":"Messages Ids","default":[]},"document_ids":{"items":{"type":"integer"},"type":"array","title":"Document Ids","default":[]}},"type":"object","required":["id","title"],"title":"ChatView"},"Document":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"bucket_name":{"type":"string","title":"Bucket Name"},"original_file_name":{"type":"string","title":"Original File Name"},"extension":{"type":"string","title":"Extension"},"file_name":{"type":"string","title":"File Name"},"user_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Id"}},"type":"object","required":["bucket_name","original_file_name","extension","file_name"],"title":"Document"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Message":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"chat_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Chat Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"reply_to":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Reply To"},"agentic":{"type":"boolean","title":"Agentic","default":false},"finished":{"type":"boolean","title":"Finished","default":true},"user_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Id"}},"type":"object","required":["content","reply_to"],"title":"Message"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserDTO":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","maxLength":50,"format":"email","title":"Email"},"disabled":{"type":"boolean","title":"Disabled","default":false}},"type":"object","required":["id"],"title":"UserDTO"},"UserRegistrationDTO":{"properties":{"email":{"type":"string","maxLength":50,"format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","title":"UserRegistrationDTO"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}